+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0.1] - 2024-10-22
+
+### Added
+
+- Added Unix timestamp support to between_dates() conditions
+- Date conditions are now timezone-aware (UTC)
+- Fixed some typing issues
+
+### Docs
+
+- Added technical per-object documentation pages
+- Updated themes and visuals
+- Added a changelog page that reflects CHANGELOG.md
+
 ## [1.2.0] - 2024-10-22
 
 ### Added